Move subprotocol enum, naming fixes
