+++ packages/portalnetwork/src/index.ts
@@ -5,3 +5,4 @@ export * from './subprotocols/history'
 export * from './wire'
 export * from './util'
 export * from './transports'
+export * from './subprotocols'
