+++ packages/portalnetwork/src/subprotocols/index.ts
@@ -1,0 +1,10 @@
+// subprotocol IDs
+export enum SubprotocolIds {
+  StateNetwork = '0x500a',
+  HistoryNetwork = '0x500b',
+  TxGossipNetwork = '0x500c',
+  HeaderGossipNetwork = '0x500d',
+  CanonicalIndicesNetwork = '0x500e',
+  UTPNetwork = '0x757470',
+  Rendezvous = '0x72656e',
+}
