+++ packages/portalnetwork/src/wire/types.ts
@@ -11,17 +11,6 @@ import {
   BitArray,
 } from '@chainsafe/ssz'
 
-// subprotocol IDs
-export enum SubprotocolIds {
-  StateNetwork = '0x500a',
-  HistoryNetwork = '0x500b',
-  TxGossipNetwork = '0x500c',
-  HeaderGossipNetwork = '0x500d',
-  CanonicalIndicesNetwork = '0x500e',
-  UTPNetwork = '0x757470',
-  Rendezvous = '0x72656e',
-}
-
 // Ping/Pong Custom Data type -- currently identical for State and History networks
 export const PingPongCustomDataType = new ContainerType({
   radius: new UintBigintType(32),
