+++ packages/portalnetwork/src/wire/utp/PortalNetworkUtp/PortalNetworkUTP.ts
@@ -2,7 +2,7 @@ import { Discv5 } from '@chainsafe/discv5'
 import { toHexString } from '@chainsafe/ssz'
 import { Debugger } from 'debug'
 import { bufferToPacket, ConnectionState, Packet, PacketType, randUint16, UtpSocket } from '..'
-import { SubprotocolIds } from '../..'
+import { SubprotocolIds } from '../../..'
 import { PortalNetwork } from '../../..'
 import { sendFinPacket } from '../Packets/PacketSenders'
 import { BasicUtp } from '../Protocol/BasicUtp'
